+++ project/demos/Face/FaceDemo.gd
@@ -1,7 +1,7 @@
 extends Node
 
-export (int, 0, 359, 2) var angular_speed_max := 120 setget set_angular_speed_max
-export (int, 0, 359, 2) var angular_accel_max := 10 setget set_angular_accel_max
+export (int, 0, 1080, 2) var angular_speed_max := 120 setget set_angular_speed_max
+export (int, 0, 2048, 2) var angular_accel_max := 10 setget set_angular_accel_max
 export (int, 0, 180, 2) var align_tolerance := 5 setget set_align_tolerance
 export (int, 0, 359, 2) var deceleration_radius := 45 setget set_deceleration_radius
 export (float, 0, 1000, 40) var player_speed := 600.0 setget set_player_speed