+++ project/demos/FollowPath/FollowPathDemo.gd
@@ -1,7 +1,7 @@
 extends Node
 
 export (float, 0, 2000, 40) var linear_speed_max := 600.0 setget set_linear_speed_max
-export (float, 0, 200, 10.0) var linear_acceleration_max := 40.0 setget set_linear_acceleration_max
+export (float, 0, 9000, 10.0) var linear_acceleration_max := 40.0 setget set_linear_acceleration_max
 export (float, 0, 100, 0.1) var arrival_tolerance := 10.0 setget set_arrival_tolerance
 export (float, 0, 500, 10) var deceleration_radius := 100.0 setget set_deceleration_radius
 export (float, 0, 5, 0.1) var predict_time := 0.3 setget set_predict_time